refactor(agent): move FilteredToolExecutor to gglib-agent to eliminate duplication
Extracts the FilteredToolExecutor decorator from gglib-axum's inline agent
handler definition into a new crates/gglib-agent/src/filter.rs module.

Re-exports as gglib_agent::FilteredToolExecutor so both gglib-axum (Phase 3)
and the upcoming gglib-cli (Phase 4) consume the single canonical
implementation — no copy-paste, no divergence risk.

The gglib-axum handler is updated to use FilteredToolExecutor::new() via the
shared import; all logic is identical, only the location changes.

Closes no issue by itself; prerequisite for #267 Phase 4 CLI implementation.
