+++ crates/gglib-agent/src/lib.rs
@@ -4,9 +4,11 @@
 
 pub mod agent_loop;
 pub mod context_pruning;
+pub mod filter;
 pub mod loop_detection;
 pub mod stagnation;
 pub mod stream_collector;
 pub mod tool_execution;
 
 pub use agent_loop::AgentLoop;
+pub use filter::FilteredToolExecutor;
